+++ tests/python/common/test_batch-graph.py
@@ -4,7 +4,7 @@ import backend as F
 
 import dgl
 import numpy as np
-from pytests_utils import parametrize_idtype
+from utils import parametrize_idtype
 
 
 def tree1(idtype):
