+++ tests/python/common/test_batch-heterograph.py
@@ -5,7 +5,7 @@ import backend as F
 import dgl
 import pytest
 from dgl.base import ALL
-from pytests_utils import check_graph_equal, get_cases, parametrize_idtype
+from utils import check_graph_equal, get_cases, parametrize_idtype
 
 
 def check_equivalence_between_heterographs(
