+++ tests/python/common/test_heterograph-remove.py
@@ -2,7 +2,7 @@ import backend as F
 
 import dgl
 import numpy as np
-from pytests_utils import parametrize_idtype
+from utils import parametrize_idtype
 
 
 @parametrize_idtype
