+++ tests/python/common/test_homophily.py
@@ -4,7 +4,7 @@ import unittest
 import backend as F
 
 import dgl
-from pytests_utils import parametrize_idtype
+from utils import parametrize_idtype
 
 
 @unittest.skipIf(
