+++ tests/python/common/test_merge.py
@@ -1,7 +1,7 @@
 import backend as F
 
 import dgl
-from pytests_utils import parametrize_idtype
+from utils import parametrize_idtype
 
 
 @parametrize_idtype
