+++ tests/python/common/test_readout.py
@@ -6,8 +6,8 @@ import dgl
 import networkx as nx
 import numpy as np
 import pytest
-from pytests_utils import parametrize_idtype
-from pytests_utils.graph_cases import get_cases
+from utils import parametrize_idtype
+from utils.graph_cases import get_cases
 
 
 @parametrize_idtype
