+++ tests/python/common/test_sparse_ops-csr.py
@@ -4,7 +4,7 @@ import dgl
 import numpy as np
 import pytest
 import scipy.sparse as ssp
-from pytests_utils import parametrize_idtype
+from utils import parametrize_idtype
 
 if F.backend_name == "pytorch":
     import torch
