+++ tests/python/common/test_subgraph.py
@@ -7,7 +7,7 @@ import networkx as nx
 import numpy as np
 import pytest
 import scipy.sparse as ssp
-from pytests_utils import parametrize_idtype
+from utils import parametrize_idtype
 
 D = 5
