+++ tests/python/common/transforms/test_to_block.py
@@ -18,7 +18,7 @@ import backend as F
 
 import dgl
 import dgl.partition
-from pytests_utils import parametrize_idtype
+from utils import parametrize_idtype
 
 
 @parametrize_idtype
