+++ tests/python/common/transforms/test_transform.py
@@ -26,8 +26,8 @@ import dgl.partition
 import networkx as nx
 import numpy as np
 import pytest
-from pytests_utils import parametrize_idtype
-from pytests_utils.graph_cases import get_cases
+from utils import parametrize_idtype
+from utils.graph_cases import get_cases
 from scipy import sparse as spsp
 
 D = 5
