+++ tests/python/common/utils/test_filter.py
@@ -5,7 +5,7 @@ import backend as F
 import dgl
 import numpy as np
 from dgl.utils import Filter
-from pytests_utils import parametrize_idtype
+from utils import parametrize_idtype
 
 
 def test_graph_filter():
