+++ tests/python/tensorflow/test_nn.py
@@ -10,8 +10,8 @@ import numpy as np
 import pytest
 import scipy as sp
 import tensorflow as tf
-from pytests_utils import parametrize_idtype
-from pytests_utils.graph_cases import (
+from utils import parametrize_idtype
+from utils.graph_cases import (
     get_cases,
     random_bipartite,
     random_dglgraph,
