+++ tests/utils/graph_cases.py
